// Frontend/src/app/shared/models/CustomerRegister.ts
export class CustomerRegister{
    FirstName: string = ''; 
    MiddleName: string = ''; 
    LastName: string = ''; 
    Phone: string = ''; 

    AddressLine1: string = ''; 
    AddressLine2: string = ''; 
    City: string = ''; 
    StateProvince: string = '';
    CountryRegion: string = ''; 
    PostalCode: string = '';

    EmailAddress: string = ''
    Password: string = ''

    AddressType: string = '';
}